+++ src/models/egli.cc
@@ -35,8 +35,8 @@ static __inline float _10log10f(float x) {
     return (4.342944f * logf(x));
 }
 
-double EgliPathLoss(float f, float h1, float h2, float d) {
-    double Lp50 = NAN;
+float EgliPathLoss(float f, float h1, float h2, float d) {
+    float Lp50 = NAN;
     float C1, C2;
 
     /*  if ((f >= fcmin) && (f <= fcmax) &&
