+++ src/models/pel.cc
@@ -16,13 +16,13 @@
 
 #include <math.h>
 
-double PlaneEarthLoss(float d, float TxH, float RxH) {
+float PlaneEarthLoss(float d, float TxH, float RxH) {
     /*
     Plane Earth Loss model
     Frequency: N/A
     Distance (km): Any
     */
     // Plane earth loss is independent of frequency.
-    double dbloss = 40 * log10(d) + 20 * log10(TxH) + 20 * log10(RxH);
+    float dbloss = 40 * log10(d) + 20 * log10(TxH) + 20 * log10(RxH);
     return dbloss;
 }
